Move isTest to different module to avoid cyclic depedendency
