+++ src/utils/tests.ts
@@ -54,7 +54,3 @@ function writeJsonFile(data: any, prefix: string): string {
     fs.writeFileSync(file.fd, json);
     return filePath;
 }
-
-export function isTest(): boolean {
-    return process.env.REACT_APP_CYPRESS === "true" || process.env.JEST_WORKER_ID !== undefined;
-}
